missing input directory fix when using terminal UI
Just a quick fix. Github won't let me upload empty directories. So if you end up running `commandUI.py` and execute on an input directory that does not exist, there will be an error.

Conclusion: User has to create input directory first, or wait until `commandUI.py` runs for it to appear.

(also added a `.gitignore` for __pycache__ when committing)