More keycount support
Added support for maps with keycounts lower than 7k. More specifically, I have added support for keycounts from 2k up to 6k. (1k presented some issues that I will try and resolve).

Created a dictionary `lt7keys` that maps the keys to a certain column in the bms 8k playfield. This allows easy formatting and easy modification if needed. If I didn't do this, it would have been a LOT of if statements. And nobody likes that.